feat: finalizo implementacion de prisma con manejo de errores

+++ src/modules/instructor/instructor.service.ts
@@ -1,5 +1,5 @@
+import { Instructor } from '../../generated/prisma/client.js';
 import { InstructorRepository } from './instructor.repository.js';
-import { InstructorProps } from './instructor.entity.js';
 import {
   CreateInstructorInput,
   UpdateInstructorInput,
@@ -8,11 +8,11 @@ import {
 export class InstructorService {
   constructor(private instructorRepository: InstructorRepository) {}
 
-  async getAll(): Promise<InstructorProps[]> {
+  async getAll(): Promise<Instructor[]> {
     return await this.instructorRepository.getAll();
   }
 
-  async getById(id: number): Promise<InstructorProps> {
+  async getById(id: number): Promise<Instructor> {
     const instructor = await this.instructorRepository.getOne(id);
     if (!instructor) {
       throw new Error('Instructor no encontrado');
@@ -20,25 +20,30 @@ export class InstructorService {
     return instructor;
   }
 
-  async add(props: CreateInstructorInput): Promise<InstructorProps> {
+  async add(props: CreateInstructorInput): Promise<Instructor> {
     const newInstructor = await this.instructorRepository.add({
       ...props,
     });
     return newInstructor;
   }
 
-  async update(
-    id: number,
-    props: UpdateInstructorInput,
-  ): Promise<InstructorProps> {
+  async update(id: number, props: UpdateInstructorInput): Promise<Instructor> {
+    const instructor = await this.instructorRepository.getOne(id);
+    if (!instructor) {
+      throw new Error('Instructor no encontrado');
+    }
     const updatedInstructor = await this.instructorRepository.update(id, props);
     if (!updatedInstructor) {
-      throw new Error('Instructor no encontrado');
+      throw new Error('Error al actualizar el instructor');
     }
     return updatedInstructor;
   }
 
-  async delete(id: number): Promise<boolean> {
-    return await this.instructorRepository.delete(id);
+  async delete(id: number): Promise<void> {
+    const instructor = await this.instructorRepository.getOne(id);
+    if (!instructor) {
+      throw new Error('Instructor no encontrado');
+    }
+    await this.instructorRepository.delete(id);
   }
 }
